refact: View Extensions 관련 리팩토링
